signal: Ensure SIGNAL_GROUP_EXIT gets set
There is a theoretical possibility that the last thread in a thread
group will exit by calling the raw exit syscall, and not exit_group.
It is highly unlikely but to keep that from being a special case
elsewhere in the code ensure that no matter how the last thread exits
we always set SIGNAL_GROUP_EXIT and populate exit_code appropriately.

Move computation of group_dead into exit_signals.  This allows the
code to know if this really is the last task to run do_exit in the
thread_group.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>