proc: Stop using lock_task_sighand
With the modest exception of collect_sigign_sigcatch none of the users
of lock_task_sighand actually care about sighand struct or if it is
alive.  In proc lock_task_sighand appears to have been used in case a
task that was being displayed was reaped while the proc code was
running and thus would have caused an oops when attempting to
dereference sighand to take siglock.

Now that siglock is no longer stored in sighand_struct it is safe
to not use lock_task_sighand from proc.  Simplifying the logic
of the code.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>