posix-timers: Factor out do_timer_delete from timer_delete and exit_itimers
Make the code more maintainable by placing the common code in a common
function, and get rid of the misnamed itimer_delete.  Nothing in this
code deals with timers that setitimer and getitimer refer to.

This fixes a small race with reading /proc/<pid>/timers by always
taking siglock over the delete from the posix_timers list.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>