posix-timers: Rename exit_itimers exit_posix_timers
Rename exit_itimers to exit_posix_timers as exit_itimers has nothing to do
with the interval timers controlled by setitimer and getitimer.  Nor does
it have anything to do with the code in kernel/time/itimer.c

As the function has everything to do with the timers in kernel/time/posix-timers.c
and as that file uses the string posix_timers to identify it's functions the
name exit_posix_timers appears appropriate.

As I am touching this function modify it's declaration in the header
file to be a noop inline function when !CONFIG_POSIX_TIMERS.  Removing
the need to call this function under CONFIG_POSIX_TIMERS.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>