posix-timers: Rename flush_itimer_signals flush_posix_timer_signals
Rename flush_itimer_signals as it has nothing to do with the signals
controlled by setitimer and getitimer and in fact it does not flush
signals generated by those itimers.

What the code does is flush queue signals sent by the the timers
created with create_timer.

As I am renaming the function add a #ifdef CONFIG_POSIX_TIMER guard in
the header so the code doesn't needed to be called in an ungly #ifdef
CONFIG_POSIX_TIMER block.  Making the code easier to maintain.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>