posix-timers: Implement itimers_exit
In do_exit remove the magic hrtimers_cancel and replace it with a function
that makes it clear that the itimers are being cleaned up.  With the process
gone prof and virt timers don't need any cleanup but the real time timer
does.

This implementation includes #ifdef CONFIG_POSIX_TIMERS boiler plate
in the appropriate header along with the declaration so that
the code doesn't have to carry #ifdef CONFIG_POSIX_TIMERS blocks.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>