signal: Don't use group_leader in setpgid
This makes the code clearer by indicating which properties
we actually care about:
  if !pid use our current thread group id.
  Is the caller in the same thread group as the found task
     or in the same thread group as the parent of the found task.

These are concepts that can be implemented with the use of group_leader
but are not as clear.

The concept of group_leader is also a bit problematic as it changes
sometimes and thus can be a bit racy.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>