signal: Give priority to SIGKILL
If we have requested that a task be killed without coredumps or
recourse to userspace handlers don't waste resources and kill it ASAP.

Today if there are several pending signals we can wind up delivering
several signals to userspace before we process the SIGKILL.  Worse
only the magic of the signal_group_exit() test in do_signal_stop
prevents a processes from stopping before processing SIGKILL.

Make the kernel faster and simpler by always processing SIGKILL
first if it is pending.  Remove the sig_kernel_exit() check
in do_signal_stop as it is no longer needed with this change.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>