binfmt_elf: Hard code the process state to running in the core dump
Depending on which thread is picked the process state is either going
to be running (because the thread that is performing the coredump is
picked) or it is going to be sleeping.

The code has gotten out of sync with the task flags and no one has noticed.
I can't possibly see a useful value to include given the coredump process
will determine the state.  Therefore just hard code that the task
is running and forget about trying to get the value correct.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>