binfmt_elf: Revert fill_prstatus times to something sensible
When thread support was originally added the times of all of the
threads would add up to the times of the process.  This was broken
during a large rewrite of the timer code and apparently no one
noticed.

I am in the process of allowing zombie thread group leaders to
be reaped.  In such an event there will be no real thread group
leader from the perspective of userspace.  As such it will be
impossible to read the core dump and figure out which thread
reports all of the time and which threads just report themselves.

Revert the logic for reporting times in a core dump so that there is
no special thread.  The result is all threads reporting just their own
time with the exception of one thread that reports it's own time plus
the time of the exited threads.  This is less work and more
comprehensible when reading the core dump.

Fixes: f06febc96ba8 ("timers: fix itimer/many thread hang")
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>