proc: Remember if an inode references a tgid or a tid
Store a pid type in struct proc_inode to remember if the
inode is for an entire thread group or just for an individual
task.  Use that type when going from proc inodes to tasks.

This is necessary to support zombie thread group leader
exiting.

Implement this by passing the type to proc_pid_make_inode,
and initializing the type explicitly or by inheritining it from
the parent inode.

Watch our step carefully so that /proc/<pid>/ continues
to work as well as /proc/<tgid>/

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>