proc timers: Simplify the logic with get_proc_task
Cache not the pid but the proc inode and use get_proc_task
on that inode to get the task the inode points at.

This removes the need for the code to know what kind of proc inode
it is dealing with, and keeps the benefits of not keeping a task
reference where threads may come and go in the task group and
the code can continue to print the timers list for the thread
group.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>