signal: Don't send signals to zombies
When working with posix threads ensure that if there is a zombie
thread (such as the thread group leader) it is not possible to take
advantage of that zombie and it's stale struct cred to bypass
permission checks for sending signals.

In userspace nptl ensures that calls such as setresuid update all
threads so they have the same credentials.  If a zombie thread remains
such a call is a problem as the zombie threads can not be updated.

Avoid sending signals to zombies by checking for PF_EXITING on a
destination task.  This is sufficient for both group signals
and thread specific signals as pid_task is now guaranteed not
to return PF_EXITING unless the thread group is completely dead.

Use this logic in the pure signal sending functions and in
do_notify_parent and do_notify_parent_cldstop to ensure not
even sighand is accessed after a PF_EXITING is set.
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>

Fixes: Linux 2.6.0 (AKA when CLONE_THREAD started meaning something)
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>