signal: Separate exec and from group exit
Move the test out of group_exec_task out of signal_group_exit so that
exec and group exit can be separated.

Not testing group_exec_task in during a coredump allows coredumps
to be triggerd in parallel with exec and to kill the exec'ing process.

Explicitly testing group_exec_task in do_group_exit makes it clear
that group_exec_task signals a multi-threaded exec not a group exit.

Testing __fatal_siganl_pending in de_thread instead of
signal_group_exit makes it clear that de_thread is not a group exit
and does not resue any of the signal group exit fields.

There are no other users of signal_group_exit.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>