signal: Update do_notify_parent_cldstop to not use group_leader
Replace doing task_pid_nr_ns(tsk->group_leader, ...) with
pid_nr_ns(task_pid_type(tsk, type), ...).

Instead of assuming which pid to use based on if the the destination
task pass in the pid type we are notifying for.  Making it clearer what
is expected of do_notify_parent_cldstop.

This changes allows the code to function properly when the code starts
allowing the first thread of a thread group to exit and be reaped,
instead of requiring it to be a zombie.

Of particular note is the special case for sending CLD_CONTINUED to
the ptracer when it is triggered.  This is the only case that sends a
tgid report to a ptracer, and is a bit of a kludge.  This change
to do_notify_parent_cldstop makes a bit less of a kludge as
games with tsk->group_leader no longer need to be played to send
the appropriate pid number.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>