signal: Add a lock to protect just the contents of the sighand structure
The sighand_struct siglock needs to be moved to signal_struct, to make
the lifetime of the lock match the lifetime of what the lock protects.
This requires adding an additional lock in sighand_struct just for
those fields that live in sighand_struct.

To keep the large change as trivial as possible introduce
sighand_struct->lock here and use it where needed.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>