posix-cpu-timers: Consolidate the clockid_t validation
Add a helper function cpu_timer_task_rcu that takes a clockid_t value,
validates the value, finds the appropriate task for the clock,
and then validates we are allowed to access that task.

With the use of PIDTYPE_TGID I am able to remove the
has_group_leader_pid test.

This rewrite also makes the validation robust in the case
where we allow the thread group leader to exit and be
reaped before the rest of the thread group is reaped.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>