posix-cpu-timers: Store a reference to a pid not a task
This takes a little more code but it reduces the worst case memory
footprint of posix cpu timers as it no longer pins a task_struct.

Furthermore tracking processes this way very cleanly handles the case
of a tracked processes calling exec from a non-leader pid.  Which will
result in a different task being the thread group leader.  Using pids
allows us to look up a new task for the thread group when needed
instead of pinning the first task for the thread group indefinitely.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>