yama: Track ptrace exceptions by tgid not by task_struct
Yama currentl makes a strong assumption that all thread groups have a
leader task and that leader tasks will outlive the rest of the tasks.
That really is not true today.

It definitely won't be true when the work to allow the first thread in
a thread group to be reaped before the thread group exits is complete.

Therefore do the sane thing and track thread groups by the struct pid
of their thread group.  In practice this represents no change to
functionality.

To make this happen yama's security hook is moved from task free to
pid free.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>