rculist: Implement hlist_replace_init_rcu
Based on list_del_init_rcu and hlist_replace_rcu.  The only code
difference is old->pprev = NULL not LIST_POISON2;

The usage difference is that now hlist_unhashed may be used to
see if an hlist_node has either never been placed in a list
or a hlist_node has been removed from the list.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>