pidns: Remove races from reboot_pid_ns
If the pid namespaces is already exiting it doesn't make sense
for reboot_pid_ns to change the exit code and confuse things.
Therefore grab the siglock and test sig->flags before
sending a SIGKILL.

Instead of replying on magic in zap_pid_ns_processes to set
group_exit_code use start_group_exit to set SIGNAL_GROUP_EXIT and
group_exit_code directly.

This is a little more code but this removes the possibility of races,
removes unnecessary code form other functions and makes the code use
a common idiom which makes it more maintainable in the long run.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>