signal: Update do_notify_parent to allow the first thread to be reaped
Make it explicit by passing in a new paramenter to do_notify_parent
that there are two types of notifications that do_notify_parent
performs a tgid level notification where the tgid is reported as ready
to reap to the processes real parent, and a pid level notification
where the pid is reported as ready to reap to it's ptracer.

Remove the now unnecessary signal parameter as the signal that is sent
sent is dependent upon which kind of notification that is being
performed.

These changes together make the function of the code a little more
obvious and make it ready for when we allow reaping of zombie
thread group leaders.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>