pidns: Use common idioms for terminating the pid namespace
Teach zap_pid_ns_processes to place it's zombies on a
list for release_task, and remove the sleep.

Add a helper "reapable" that says ordinary processes are reapable when
all of the threads have exited and been reaped, and for child reapers
also requires that the pid hash table contains just the pids
for the child reaper.  Use reapable where thread_group_empty
was previously used to ensure processes are not reaped too soon.

Move the call of acct_exit_ns into release_task and only call
it when the last task of a pid namespace init is being reaped.
Not perfect but the best that is doable without introducing
a per pid namespace count like signal->live that would let
do_exit know when the last task in the pid namespace is
passing through it.

As the init process now terminates, replace a wake up of the init
process from free_pid by a notification of init's parent when every
task in the pid namespace except the first task of the init has
been reaped.

The result is code that looks like the rest of the exit
path in linux and as such code that should be much easier
to maintain.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>