exit: In __exit_signal directly test if this is the last thread
__exit_signal knows that __unhash_process will under the siglock
remove tsk from the list of threads.  Use that knowledge and
before the removal test if the list only has a single element,
to see if this is the last task of a thread group being reaped.

This construct makes it clearer what the code is trying to do,
making it easier to understadand and more robust to future updates.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>