exec: Don't wait for threads to be reaped.
Ever since CLONE_THREAD support was added to the kernel it has been
possible to dead-lock userspace by ptracing a process and not reaping
it's child threads.  With the use of the cred_guard_mutex in proc the ways
userspace can unknowningly trigger a dead-lock have grown.  Which makes
a simple -f strace on a program that performs a mult-threaded exec
unsafe.

        #include <unistd.h>
        #include <signal.h>
        #include <sys/ptrace.h>
        #include <pthread.h>

        void *thread(void *arg)
        {
                ptrace(PTRACE_TRACEME, 0,0,0);
                return NULL;
        }

        int main(void)
        {
                int pid = fork();

                if (!pid) {
                        pthread_t pt;
                        pthread_create(&pt, NULL, thread, NULL);
                        pthread_join(pt, NULL);
                        execlp("echo", "echo", "passed", NULL);
                }

                sleep(1);
                // or anything else which needs ->cred_guard_mutex,
                // say open(/proc/$pid/mem)
                ptrace(PTRACE_ATTACH, pid, 0,0);
                kill(pid, SIGCONT);

                return 0;
        }

Sovle this by modifying exec to only wait until all of the other
threads are zombies, and not waiting until the other threads
are reaped.

As well as simplifying the userspace semantics this simplifies
the code.

During the development of CLONE_THREAD support a lot of cleanup code
landed in release_task and __exit_signal as it was one one place that
knew when a thread group had exited.  In practice this resulted in a
lot of clean up code being called much later than necessary.  This
change had to wait until all of the interesting cleanup code was moved
out of release_task and into do_exit.  This code movement has
made waiting for a zombie sufficient for correct operation of exec.

Reported-by: Ulrich Obergfell <uobergfe@redhat.com>
Reported-by: Attila Fazekas <afazekas@redhat.com>
Reported-by: Aleksa Sarai <asarai@suse.com>
Reported-by: Oleg Nesterov <oleg@redhat.com>
Historical Tree: https://git.kernel.org/pub/scm/linux/kernel/git/tglx/history.git
Fixes: 63540ceac6d7 ("[PATCH] thread-exec-2.5.34-B1, BK-curr")
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>