exit: Move add_device_randomness to finish_task_switch
No locking is needed as the task in question is guaranteed to never
run again.

finish_task_switch is where this update makes sense.  The quantity
sum_exec_runtime that is being mixed into random number generator
state is maintained by the scheduler and is finalized after a process
finishes execing.  Further add_device_randomness is a comparatively
small function with that uses very few stack variables.  All of which
point to finish_task_switch as the perfect place to call
add_device_randomness with sum_exec_time.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>