exit: Move __cleanup_sighand from __exit_signal to exit_signals
Cleaning up in sighand struct has had to live in __exit_signal as the
sighand struct used to hold siglock and signals were previously
delivered to zombie threads that had the pid of the thread group.  Now
that we no longer do that it is no longer necessary to cleanup
the sighand struct from __exit_signal and we can move the code into
exit_signals with cleanup of the rest of signal handling.

Further this is needed to allow the first thread of a thread group
to be reaped.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>