wait: Support reaping the first thread
The pid returned from and passed into wait can refer to two distinct
kinds of objects.  These pids can refer to a thread when it is being
ptraced, and otherwise to a thread group.  Further there is an ambiguity
possible when the first thread of a child thread group is also ptraced.

The ambiguity in wait is resolved in terms of a reporting thread group level
events that are a superset of the thread level events.

Utilize the recently added PIDTYPE_TGID to allow disambiguating our filter
pid at the start.  If the pid refers to a thread group and the process is a
child use PIDTYPE_TGID otherwise if the pid refers to a thread and the
tread is a ptraced child use PIDTYPE_PID.  This allows the rest of the wait
code to know exactly what is being waited for.

The use of PIDTYPE_TGID ensures that when the first thread exits
the thread group will be found even though though there is no longer
a thread with the specified pid.

When reporting pids of traced threads report the pid.  When reporting
pids of entire thread groups report the tgid.  This allows programs
to see what they are expecting, and ensures that there is no
visible change in behavior when the code starts allowing the
first thread in a group to exit.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>