wait: Add wait_pid_task returning the child list member wait for PIDTYPE_TGID
The function do_wait_pid needs not an arbitary living thread of a thread group
but the member of the thread group that is on the wait list.  Adding a wrapper
around pid_task that returns that member, and update do_wait_pid.

Use wait_pid_task in exec instead of group_leader.  This is safe
without locks as none of "task_tgid(tsk)",  "pid_task(pid, PIDTYPE_TGID)" or
"list_first_entry(...->thread_head, ...)" changes so long as a signle
task is kept alive.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>