exit: Allow the first thread to exit
The first thread sticking around as a zombie has caused all kinds
of problems in the code and in userspace.

Allowing the first thread to exit ensures that code does not need
special cases to handle a zombie thread group leader.  Zombie thread
group leaders almost always resulted in larger code to handle the
special case or bugs.

When a multi-threaded application calls setresuid or any of it's
cousins a zombie thread (for obvious reasons) won't have it's
credentials updated.  This has proven to be a problem for things such
as the ptrace_may_access check which allowed access to the state of an
entire process based on stale credentials.  This code is now fixed but
if there are any other places where stale credentials matter allowing
the first thread to exit will make the vulnerable window much smaller.

Unfortunately when a processes is ptracing it's child there is an
ambiguity between the pid of the ptraced child and the tgid
of the thread group in which case the zombie thread must stay
around to guarantee that the pids used by wait and the siginfo
of SIGCHLD are not ambiguous.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>