signal: Move siglock from sighand_struct to signal_struct
The use of siglock in sighand_struct to lock all fields in
signal_struct as well as sighand_struct while it slightly simplifies
the locking during signal delivery and when setting signal handlers,
it significantly increases the complexity of taking the lock
everywhere else siglock is used.

The most painful situation is when we are looking up and locking the
thread group leader.  Without this change the only form that avoids
all races is a loop as it is theoretically possible for the thread group
leaders to keep dying and passing the baton (via exec today) to a new
thread group leader.

There is already a loop of just this form in kill_pid_info and several
other kill function are susceptible to subtle races they lack such a loop.

The locking complexity is also painful in de_thread where we are
required to wait until threads are reaped before allowing exec to
continue.  As exec might change the sighand_struct and __exit_signals
uses siglock to sighand_struct to access all kinds of fields when
reaping a process.

Placing the siglock in signal_struct removes all of those concerns.
It does require a new lock to be added to sighand_struct that
protects just the array that describes what action should
be taken when the kernel delivers a signal.  The new sighand->lock
nests inside of siglock.

With the movement of siglock out of sighaand_struct it is possible to
trivially take siglock without any loops, which simplifies
lock_task_siglock.

The function lock_task_siglock is retained only because a NULL
tsk->sighand pointer serves as flag to report a task that has been
reaped.  Many of the callers of lock_task_siglock depend on not
running on reaped tasks.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>